requirements.txt
# PDF Processing
PyMuPDF>=1.24.0
pdfplumber>=0.11.0
Pillow>=10.0.0  # For image processing in VLM

# LLM Integration
openai>=1.50.0
anthropic>=0.30.0
google-generativeai>=0.3.0  # Google Gemini
huggingface-hub>=0.20.0  # Hugging Face Inference API
requests>=2.31.0  # For Ollama local API calls

# Search APIs
duckduckgo-search>=6.0.0
tavily-python>=0.5.0

# State Management
redis>=5.1.0

# NLP & Text Processing
spacy>=3.8.0

# Utilities
python-dotenv>=1.0.1
pydantic>=2.9.0
httpx>=0.27.0
tenacity>=8.5.0

# Development
pytest>=8.0.0
black>=24.0.0

pyyaml>=6.0.0
colorlog>=6.10.0

reportlab >= 4.4.9